feat: limitando o acesso as rotas pelas roles

+++ build/http/controllers/courses/destroy.js
@@ -24,10 +24,10 @@ __export(destroy_exports, {
 });
 module.exports = __toCommonJS(destroy_exports);
 
-// src/use-cases/errors/course-already-exists-error.ts
-var CourseAlreadyExistsError = class extends Error {
+// src/use-cases/errors/course-not-found.ts
+var CourseNotFoundError = class extends Error {
   constructor() {
-    super("Course name already exists.");
+    super("Course not found.");
   }
 };
 
@@ -38,11 +38,25 @@ var ResourceNotFoundError = class extends Error {
   }
 };
 
+// src/env/index.ts
+var import_config = require("dotenv/config");
+var import_zod = require("zod");
+var envSchema = import_zod.z.object({
+  NODE_ENV: import_zod.z.enum(["dev", "test", "production"]).default("dev"),
+  JWT_SECRET: import_zod.z.string().optional(),
+  PORT: import_zod.z.coerce.number().default(3333)
+});
+var _env = envSchema.safeParse(process.env);
+if (_env.success === false) {
+  console.error("Invalid environment variables", _env.error.format());
+  throw new Error("Invalid environment variables.");
+}
+var env = _env.data;
+
 // src/lib/prisma.ts
 var import_client = require("@prisma/client");
 var prisma = new import_client.PrismaClient({
-  // log: env.NODE_ENV === 'dev' ? ['query', 'info', 'warn', 'error'] : [],
-  log: ["query", "info", "warn", "error"]
+  log: env.NODE_ENV === "dev" ? ["query", "info", "warn", "error"] : []
 });
 
 // src/repositories/prisma/prisma-course-repository.ts
@@ -103,6 +117,10 @@ var DestroyCourseUseCase = class {
   async execute({
     id
   }) {
+    const findCourse = await this.coursesRepository.findById(id);
+    if (!findCourse) {
+      throw new CourseNotFoundError();
+    }
     return await this.coursesRepository.destroy(
       id
     );
@@ -117,10 +135,10 @@ function makeDestroyCourseUseCase() {
 }
 
 // src/http/controllers/courses/destroy.ts
-var import_zod = require("zod");
+var import_zod2 = require("zod");
 async function destroy(request, reply) {
-  const registerBodySchema = import_zod.z.object({
-    id: import_zod.z.coerce.number()
+  const registerBodySchema = import_zod2.z.object({
+    id: import_zod2.z.coerce.number()
   });
   const { id } = registerBodySchema.parse(request.params);
   try {
@@ -129,7 +147,7 @@ async function destroy(request, reply) {
       id
     });
   } catch (err) {
-    if (err instanceof CourseAlreadyExistsError) {
+    if (err instanceof CourseNotFoundError) {
       return reply.status(409).send({ message: err.message });
     }
     if (err instanceof ResourceNotFoundError) {
